Fragment, LayerBoxFooter, Remove reduntant calls on handler, renamed updateVisibility to updateLayerVisibility, add check isInScale to geometryMatch
